+++ my-school/src/Components/PortfolioLog.js
@@ -4,6 +4,7 @@ import {
   getAllActivitiesForUser
 } from '../actions/actions-portfolio.js';
 import ActivityCard from './ActivityCard';
+import Labels from './Labels';
 import '../App.css';
 import ReactGA from "react-ga";
 import Loader from "react-spinners/ClimbingBoxLoader";
@@ -31,8 +32,9 @@ const PortfolioLog = ({activities, getAllActivitiesForUser, isLoading}) => {
       }, [activities])
     return(
         <div className='portfolio-list'>
+          <Labels />
           {isLoading === true ? <Loader color={'#329795'} css={override}/> : 
-          (sortedActivities.map(activity =>(<ActivityCard key={activity.id} activity={activity} className='card' />))) }
+          (sortedActivities.map(activity =>(<ActivityCard key={activity.id} activity={activity} className='card' />)))}
         </div>
     )
 };
